Added folllowing enemy

+++ src/com/jodelahithit/main/HUD.java
@@ -5,8 +5,8 @@ import java.awt.Graphics;
 
 public class HUD {
 
-	public static int HEALTH = 100;
-	private int greenValue = 255, redValue = 0;
+	public static float HEALTH = 100;
+	private float greenValue = 255, redValue = 0;
 	
 	private int score = 0;
 	private int level = 1;
@@ -39,8 +39,8 @@ public class HUD {
 	public void render(Graphics g){
 		g.setColor(Color.gray);
 		g.fillRect(15, 15, 200, 32);
-		g.setColor(new Color(redValue, greenValue, 0));
-		g.fillRect(15, 15, HEALTH*2, 32);
+		g.setColor(new Color((int)redValue, (int)greenValue, 0));
+		g.fillRect(15, 15, (int)HEALTH*2, 32);
 		g.setColor(Color.white);
 		g.drawRect(15, 15, 200, 32);
