+++ src/com/jodelahithit/main/objects/BasicEnemy.java
@@ -13,7 +13,7 @@ public class BasicEnemy extends GameObject {
 
 	private Handler handler;
 
-	public BasicEnemy(int x, int y, ID id, Handler handler) {
+	public BasicEnemy(float x, float y, ID id, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		velX = 5;
@@ -34,10 +34,10 @@ public class BasicEnemy extends GameObject {
 
 	public void render(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 }
